Implement `DataFrameBinOp`
